fix(widget-android): inline rendering — drop ListView/RemoteViewsService
Samsung One UI launchers fail to bind RemoteViewsAdapter, leaving the
widget stuck on the system "Chargement…" placeholder. Render the up-to-5
articles directly in onUpdate() via RemoteViews.addView into a
LinearLayout container, eliminating the Service/Factory pair.

Wordmark now uses system serif bold (Fraunces was being stripped by the
launcher); subtitle bumped to 13sp.
