--- apps/mobile/android/app/src/main/res/layout/widget_loading_view.xml
@@ -1,25 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!-- Skeleton row shown by the ListView while RemoteViewsFactory is loading. -->
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:orientation="vertical"
-    android:padding="8dp">
-
-    <View
-        android:layout_width="80dp"
-        android:layout_height="10dp"
-        android:layout_marginBottom="6dp"
-        android:background="@color/facteur_border" />
-
-    <View
-        android:layout_width="match_parent"
-        android:layout_height="14dp"
-        android:layout_marginBottom="4dp"
-        android:background="@color/facteur_border" />
-
-    <View
-        android:layout_width="180dp"
-        android:layout_height="14dp"
-        android:background="@color/facteur_border" />
-</LinearLayout>
